move footer to container folder
