+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import './base.css';
 import './modules.css';
-import Footer from './components/atoms/footer';
+import Footer from './components/templates/container/Footer';
 
 class App extends Component {
   render() {
